rename user to users to avoid admin conflict, add admin
