+++ purchasing/extensions.py
@@ -22,4 +22,4 @@ from flask_debugtoolbar import DebugToolbarExtension
 debug_toolbar = DebugToolbarExtension()
 
 from flask_admin import Admin
-admin = Admin()
+admin = Admin(endpoint='admin')
